+++ tests/test_project.py
@@ -6,6 +6,7 @@ import textwrap
 
 import pytest
 
+import pyflp
 from pyflp.exceptions import ExpectedValue
 from pyflp.project import VALID_PPQS, FileFormat, FLVersion, PanLaw, Project
 
@@ -57,3 +58,11 @@ def test_project(project: Project):
 
     with pytest.raises(ExpectedValue, match="major.minor.build.patch?"):
         project.version = "2.2"
+
+
+def test_null_check(project: Project, tmp_path: pathlib.Path):
+    pyflp.save(project, tmp_path / "null_check.flp")
+    b1 = open(pathlib.Path(__file__).parent / "assets" / "FL 20.8.4.flp", "rb").read()
+    b2 = open(tmp_path / "null_check.flp", "rb").read()
+    result = b1 == b2  # ! Don't compare 2 bytes object in pytest EVER
+    assert result